Fixed nans occurring for maxent example.
IMPORTANT: This occurred due to unsafe divisions in unscale and conicDivision in the cone module.

Fixed this for LP cones by introducing safe divisions (regularizing by EPS).

Introducing this for other divisions, too, is necessary - but care has to be taken to not alter convergence, so I have not implement this yet at other places.