Fixed issue 38 by adding static regularization to the cones, too.
Added new unit test structure under directory /test - type 'make test' to
build the executable ecostester. This one runs all tests until a problem is
detected.

So far only 1 test resides there but the structure is in place so we can start
adding more test.

The Matlab script matlab/generateTest takes an ECOS problem and generates the
required C-files, making it very easy to add more tests to the library. After
generating a test, edit test/ecostester.c and add the newly generated header
file as well as a call to the respective test such that it is included in
the test suite.